add types file for layout and task, add request file for task -- seperation of concerns
