+++ src/components/Layout/Layout.tsx
@@ -1,15 +1,14 @@
 import { useEffect, useState } from "react";
 import { TaskItem } from "../TaskItem/TaskItem";
 import { AddTaskButton, MainWrapper } from "./Layout.style";
-import { taskType } from "../TaskItem/TaskItem";
+import { TaskArray } from "./Layout.types";
 
 export const Layout = () => {
 
-  const [tasksArray, setTaskArray] = useState<Array<{ id: number, priority: string, title: string, due_date: string, completed: boolean }>>([])
+  const [tasksArray, setTaskArray] = useState<TaskArray[]>([])
 
   const handleDelete = (id: number) => {
     const filteredTasksArray = tasksArray.filter((singleTask) => {
-      console.log(singleTask)
       return singleTask.id !== id
     })
     setTaskArray(filteredTasksArray)
