+++ src/components/Layout/Layout.types.tsx
@@ -1,0 +1,9 @@
+import { priorityType } from "../TaskItem/Taskitem.types"
+
+export type TaskArray = {
+    id: number,
+    priority: priorityType,
+    title: string,
+    due_date: string,
+    completed: boolean
+}
